# Queue-based client for loki.write