--- a/component/common/loki/client/batch.go
+++ b/component/common/loki/client/batch.go
@@ -29,9 +29,6 @@ type batch struct {
 	createdAt time.Time
 
 	maxStreams int
-
-	// segmentCounter tracks the amount of entries for each segment present in this batch.
-	segmentCounter map[int]int
 }
 
 func newBatch(maxStreams int, entries ...loki.Entry) *batch {
@@ -74,14 +71,6 @@ func (b *batch) add(entry loki.Entry) error {
 	return nil
 }
 
-func (b *batch) countForSegment(segment int) {
-	if curr, ok := b.segmentCounter[segment]; ok {
-		b.segmentCounter[segment] = curr + 1
-		return
-	}
-	b.segmentCounter[segment] = 1
-}
-
 // add an entry to the batch
 func (b *batch) addFromWAL(lbs model.LabelSet, entry logproto.Entry, segment int) error {
 	b.bytes += len(entry.Line)
@@ -90,7 +79,6 @@ func (b *batch) addFromWAL(lbs model.LabelSet, entry logproto.Entry, segment int
 	labels := labelsMapToString(lbs, ReservedLabelTenantID)
 	if stream, ok := b.streams[labels]; ok {
 		stream.Entries = append(stream.Entries, entry)
-		b.countForSegment(segment)
 		return nil
 	}
 
@@ -104,7 +92,6 @@ func (b *batch) addFromWAL(lbs model.LabelSet, entry logproto.Entry, segment int
 		Labels:  labels,
 		Entries: []logproto.Entry{entry},
 	}
-	b.countForSegment(segment)
 
 	return nil
 }