--- a/component/common/loki/client/client_writeto.go
+++ b/component/common/loki/client/client_writeto.go
@@ -59,7 +59,7 @@ func (c *clientWriteTo) SeriesReset(segmentNum int) {
 	}
 }
 
-func (c *clientWriteTo) AppendEntries(entries wal.RefEntries, segment int) error {
+func (c *clientWriteTo) AppendEntries(entries wal.RefEntries, _ int) error {
 	var entry loki.Entry
 	c.seriesLock.RLock()
 	l, ok := c.series[entries.Ref]