--- a/component/common/loki/client/manager.go
+++ b/component/common/loki/client/manager.go
@@ -39,6 +39,11 @@ type Stoppable interface {
 	Stop()
 }
 
+type StoppableClient interface {
+	Stop()
+	StopNow()
+}
+
 // Manager manages remote write client instantiation, and connects the related components to orchestrate the flow of loki.Entry
 // from the scrape targets, to the remote write clients themselves.
 //
@@ -49,6 +54,9 @@ type Manager struct {
 	name        string
 	clients     []Client
 	walWatchers []Stoppable
+
+	// stoppableClients is kept separate from clients for avoiding having to couple queueClient to the Client interface
+	stoppableClients []StoppableClient
 
 	entries chan loki.Entry
 	once    sync.Once
@@ -69,20 +77,19 @@ func NewManager(metrics *Metrics, logger log.Logger, limits limit.Config, reg pr
 	clientsCheck := make(map[string]struct{})
 	clients := make([]Client, 0, len(clientCfgs))
 	watchers := make([]Stoppable, 0, len(clientCfgs))
+	stoppableClients := make([]StoppableClient, 0, len(clientCfgs))
 	for _, cfg := range clientCfgs {
 		// Don't allow duplicate clients, we have client specific metrics that need at least one unique label value (name).
-		configDigest := GetClientName(cfg)
-		if _, ok := clientsCheck[configDigest]; ok {
+		clientName := GetClientName(cfg)
+		if _, ok := clientsCheck[clientName]; ok {
 			return nil, fmt.Errorf("duplicate client configs are not allowed, found duplicate for name: %s", cfg.Name)
 		}
 
-		clientsCheck[configDigest] = fake
+		clientsCheck[clientName] = fake
 
 		if walCfg.Enabled {
-			configName := GetClientName(cfg)
-
 			// add some context information for the logger the watcher uses
-			wlog := log.With(logger, "client", configName)
+			wlog := log.With(logger, "client", clientName)
 
 			queue, err := NewQueue(metrics, cfg, limits.MaxStreams, limits.MaxLineSize.Val(), limits.MaxLineSizeTruncate, logger, QueueConfig{
 				DrainTimeout: time.Minute * 5,
@@ -91,15 +98,17 @@ func NewManager(metrics *Metrics, logger log.Logger, limits limit.Config, reg pr
 			if err != nil {
 				return nil, fmt.Errorf("error starting queue client: %w", err)
 			}
+			stoppableClients = append(stoppableClients, queue)
+
 			// subscribe watcher's wal.WriteTo to writer events. This will make the writer trigger the cleanup of the wal.WriteTo
 			// series cache whenever a segment is deleted.
 			notifier.SubscribeCleanup(queue)
 
-			watcher := wal.NewWatcher(walCfg.Dir, configName, watcherMetrics, queue, wlog, walCfg.WatchConfig)
+			watcher := wal.NewWatcher(walCfg.Dir, clientName, watcherMetrics, queue, wlog, walCfg.WatchConfig)
 			// subscribe watcher to wal write events
 			notifier.SubscribeWrite(watcher)
 
-			level.Debug(logger).Log("msg", "starting WAL watcher for client", "client", configName)
+			level.Debug(logger).Log("msg", "starting WAL watcher for client", "client", clientName)
 			watcher.Start()
 
 			watchers = append(watchers, watcher)
@@ -110,6 +119,7 @@ func NewManager(metrics *Metrics, logger log.Logger, limits limit.Config, reg pr
 			}
 
 			clients = append(clients, client)
+			stoppableClients = append(stoppableClients, client)
 		}
 	}
 	manager := &Manager{
@@ -118,10 +128,10 @@ func NewManager(metrics *Metrics, logger log.Logger, limits limit.Config, reg pr
 		entries:     make(chan loki.Entry),
 	}
 	if walCfg.Enabled {
-		manager.name = "wal"
+		manager.name = buildManagerName("wal", clientCfgs...)
 		manager.startWithConsume()
 	} else {
-		manager.name = "multi"
+		manager.name = buildManagerName("multi", clientCfgs...)
 		manager.startWithForward()
 	}
 	return manager, nil
@@ -158,22 +168,13 @@ func (m *Manager) startWithForward() {
 }
 
 func (m *Manager) StopNow() {
-	for _, c := range m.clients {
+	for _, c := range m.stoppableClients {
 		c.StopNow()
 	}
 }
 
 func (m *Manager) Name() string {
-	var sb strings.Builder
-	sb.WriteString(m.name)
-	sb.WriteString(":")
-	for i, c := range m.clients {
-		sb.WriteString(c.Name())
-		if i != len(m.clients)-1 {
-			sb.WriteString(",")
-		}
-	}
-	return sb.String()
+	return m.name
 }
 
 func (m *Manager) Chan() chan<- loki.Entry {
@@ -189,14 +190,17 @@ func (m *Manager) Stop() {
 		walWatcher.Stop()
 	}
 	// close clients
-	for _, c := range m.clients {
+	for _, c := range m.stoppableClients {
 		c.Stop()
 	}
 }
 
-// GetClientName computes the specific name for each client config. Since the name is a hash of the config as whole,
-// this allows us to detect repeated configs.
+// GetClientName computes the specific name for each client config. The name is either the configured Name setting in Config,
+// or a hash of the config as whole, this allows us to detect repeated configs.
 func GetClientName(cfg Config) string {
+	if cfg.Name != "" {
+		return cfg.Name
+	}
 	return asSha256(cfg)
 }
 
@@ -206,4 +210,18 @@ func asSha256(o interface{}) string {
 
 	temp := fmt.Sprintf("%x", h.Sum(nil))
 	return temp[:6]
+}
+
+// buildManagerName assembles the Manager's name from all configs, and a given prefix.
+func buildManagerName(prefix string, cfgs ...Config) string {
+	var sb strings.Builder
+	sb.WriteString(prefix)
+	sb.WriteString(":")
+	for i, c := range cfgs {
+		sb.WriteString(GetClientName(c))
+		if i != len(cfgs)-1 {
+			sb.WriteString(",")
+		}
+	}
+	return sb.String()
 }