#### PR Description

Right now, the `loki.write` under the hood uses loki remote write, which has two modes: WAL enabled or disabled. WAL needs some work (better segment cleanup, replay and a better batches sending mechanism), however, this effort is **not easy since the both modes share the same remote write client**.

The main purpose of this PR / effort is to have two different remote-write clients, one specific for WAL, and the existing one for non WAL. This will allow us to easily implement the following upcoming changes:
- a marker based old segments cleanup mechanism
- WAL replay
- better send semantics, using something like the queue based metrics remote write
- while not affecting the non-WAL-enabled component, which is the default and non-experimental use case

The diagrams below show, first, the current remote write implementation for when WAL is enabled.

![image](https://github.com/grafana/agent/assets/2617411/20dc2b5d-5700-4c42-a907-755128630339)

The proposed change introduces a "queueing" based client. Fist, by making the remote write client implement [`wal.WriteTo`](https://github.com/grafana/agent/blob/main/component/common/loki/wal/watcher.go#L52), WAL-specific information bleeds into the client layer, making it easy to track from which WAL segment an entry belongs to. This allows us to easily track when a segment's entries are batched, and then sent, so we can more intelligently keep track of consumed WAL segments. 
Also, right now the same remote write client routine does different tasks: builds segments, checks wether they are ready to be sent, encodes them and sends them. This should be split apart. By introducing a queue in the middle, with buffering, that holds batches ready to be sent, we can separate the batch building from the actual send operation into two different routines. Also, in the future, this allows us to easily add more queues, mimicking the `QueueManager` metrics remote write implementation.

![image](https://github.com/grafana/agent/assets/2617411/f3881884-079c-4186-b3ea-78f5637681fd)

The idea of this PR (1 in the list below) is to work as the stepping stone for:
1. PR that introduces the queue-based client
2. PR that implement the marker based segment cleanup, and WAL replay https://github.com/grafana/agent/pull/5590
3. Better send semantics through better queueing

To note there's no significant performance difference between the original client, and the queue client, the following benchmark was run writing 100 log entries with similar labels:

```
➜  client git:(pablo/wal-replay) benchstat ~/work/benches/queue-client-1.txt ~/work/benches/simple-client-1.txt
/Users/pablo/work/benches/queue-client-1.txt:4: missing iteration count
/Users/pablo/work/benches/simple-client-1.txt:4: missing iteration count
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/component/common/loki/client
               │ /Users/pablo/work/benches/queue-client-1.txt │ /Users/pablo/work/benches/simple-client-1.txt │
               │                    sec/op                    │            sec/op             vs base         │
QueueClient-10                                     1.017 ± 0%                     1.017 ± 0%  ~ (p=0.485 n=6)
```

#### Which issue(s) this PR fixes

<!-- Uncomment the following line if you want that GitHub issue gets automatically closed after merging the PR -->
<!-- Fixes #issue_id -->

#### Notes to the Reviewer

#### PR Checklist

- [x] CHANGELOG.md updated
- [x] Documentation added
- [x] Tests updated
- [ ] Config converters updated
- [x] Add benchmark