@thepalbi 2023-10-25
@mattdurham added a benchmark (the one in queue_client_test.go), which I ran with the queue client and the old one. Results here:
```
➜  client git:(pablo/wal-replay) benchstat ~/work/benches/queue-client-1.txt ~/work/benches/simple-client-1.txt
/Users/pablo/work/benches/queue-client-1.txt:4: missing iteration count
/Users/pablo/work/benches/simple-client-1.txt:4: missing iteration count
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/component/common/loki/client
               │ /Users/pablo/work/benches/queue-client-1.txt │ /Users/pablo/work/benches/simple-client-1.txt │
               │                    sec/op                    │            sec/op             vs base         │
QueueClient-10                                     1.017 ± 0%                     1.017 ± 0%  ~ (p=0.485 n=6)
```

Outcome, no significant difference. Mainly because I think the improvements in this PR will help in worst case scenarios with multi tenancy, and failure scenarios on the remote write destination.