@thepalbi 2023-10-25
@ptodev  responding inline:

> Thank you for this change! It feels like a step in the right direction. I have a few questions:
> 
> * If the Agent config gets reloaded and we switch the `wal` block from `enabled = true` to `enabled = false` (or vice versa), I guess there are no issues?

No issues, but the following would happen:
1. the manager would be stopped, stopping and draining the queue client
2. the non wal client would be enabled

The only thing I can imagine would happen here, is if the watcher is not completely over the top of the current segment, some data would be missed in that segment. But since turning on and off WAL is not an expected use case, I think we are good with that behaviour.  Regardless, with https://github.com/grafana/agent/pull/5590 WAL would be replayed and this could be handled ok.

> * I suppose the previous implementation was more synchronous? Hence, if the channel in the new implementation is full, there are no extra delays and  having the channel can only help?

Exactly, the previous implementation was all unbuffered channels connected together, therefore, synchronous. Having a buffered channel here allows us to add more asyncronicity which unblock the WAL reader end from the remote write client.

> * I guess the idea behind the "send batch work pool" is to have something similar to the Prometheus remote write sharding, except that here we would shard based on log streams?

Yes, the idea is on the third iteration mentioned  in the PR description (after wal replay is implemented, which is the other missing piece), implement something like sharding. Now, if the sharding would be done based on log streams, tenants, or a combination of both is to be decided. I think a combination of both makes sense, but we can leave that for later.