@thepalbi 2023-10-25
@mattdurham @ptodev added more test cases (different num of entries sent, with our without batching). Also added cases to the benchmark, with default batching settings and 100k entries, and a small test case, and re-run with `benchmem`:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/component/common/loki/client
                                                                    │   queue    │             regular              │
                                                                    │   sec/op   │   sec/op    vs base              │
ClientImplementations/100k_entries,_100_series,_default_batching-10   4.434 ± 2%   4.716 ± 6%  +6.36% (p=0.004 n=6)
ClientImplementations/100_entries,_single_series,_no_batching-10      1.001 ± 0%   1.016 ± 0%  +1.50% (p=0.002 n=6)
geomean                                                               2.107        2.189       +3.90%

                                                                    │    queue     │              regular               │
                                                                    │     B/op     │     B/op      vs base              │
ClientImplementations/100k_entries,_100_series,_default_batching-10   771.9Mi ± 0%   760.1Mi ± 0%  -1.53% (p=0.002 n=6)
ClientImplementations/100_entries,_single_series,_no_batching-10      842.7Ki ± 5%   830.0Ki ± 2%       ~ (p=0.058 n=6)
geomean                                                               25.20Mi        24.82Mi       -1.52%

                                                                    │    queue    │              regular              │
                                                                    │  allocs/op  │  allocs/op   vs base              │
ClientImplementations/100k_entries,_100_series,_default_batching-10   10.80M ± 0%   10.50M ± 0%  -2.77% (p=0.002 n=6)
ClientImplementations/100_entries,_single_series,_no_batching-10      11.25k ± 0%   10.94k ± 0%  -2.75% (p=0.002 n=6)
geomean                                                               348.7k        339.0k       -2.76%
```

I think at little memory expense (with the 100MiB capacity setting) we get some less CPU stress, and better semantics for write failure scenarios considering we have the buffered send queue.