@mattdurham 2023-10-26
Looking good the only comment I have is on the default, I would rather a lower default memory usage. In the case the memory buffer fills up ogs are still written to the wal and once the batches are enqueued in the case of a network issue then logs continue going to the disk until the outage is resolved?