@thepalbi 2023-10-26
@mattdurham 
> In the case the memory buffer fills up ogs are still written to the wal and once the batches are enqueued in the case of a network issue then logs continue going to the disk until the outage is resolved

Yes, since the wal writer is completely decoupled from the sending end, it will keep writing into wal even though the sender end is completely blocked.